Move db.js to a common directory, as migrate will use it as well
