+++ src/glyphs.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const db = require('./db');
+const db = require('../common/db');
 
 router.get('/', (req, res) => {
   return db.collection.find({}).sort({ layer: 1 }).toArray().then(result => (
